requirements.txt
# CLI framework
click>=8.1.0

# Testing framework
pytest>=8.3.0
pytest-html>=4.0.0
pytest-timeout>=2.2.0

# HTTP client
requests>=2.31.0
httpx>=0.27.0

# Configuration and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# GitHub integration
PyGithub>=2.1.0

# Templates and reports
jinja2>=3.1.0

# YAML parsing
pyyaml>=6.0

# Load testing (optional)
locust>=2.20.0

# Code quality
ruff>=0.8.0
black>=24.0.0
